batch: Isolate replacement edit location proofs
Replacement planning verifies candidate removal spans against saved
origins, trusted targets, and mapped source neighbors.

Those location proofs share one module with group traversal and edit
recording, making their placement rules difficult to review independently.

This commit moves the individual location proofs into a dedicated module
without changing their bodies. That boundary isolates placement evidence
before separating group verification.
